flasksrc/requirements.txt
aiohttp-jinja2==1.1.0
aiohttp-session==2.7.0
aiohttp==3.5.3
aiopg==0.15.0
aioredis==1.2.0
async-timeout==3.0.1      # via aiohttp, aioredis
attrs==18.2.0             # via aiohttp
chardet==3.0.4            # via aiohttp
hiredis==0.3.1            # via aioredis
idna==2.8                 # via yarl
jinja2==2.10              # via aiohttp-jinja2
markupsafe==1.1.0         # via jinja2
multidict==4.5.2          # via aiohttp, yarl

pyyaml==3.13
trafaret-config==2.0.2
trafaret==1.2.0
yarl==1.3.0               # via aiohttp
requests==2.25.1
Flask==1.0.3
mysql-connector-python==8.0.26
mysql-connector-python-rf==2.2.2
psycopg2==2.9.1
psutil==5.8.0
PyYAML==5.3.1
drf-spectacular==0.19.0
uwsgi==2.0.19.1
